Add variable speed based on elevator height in DriveWithJoysticks

+++ robot/src/org/usfirst/frc/team1089/robot/subsystems/Elevator.java
@@ -129,7 +129,7 @@ public class Elevator extends Subsystem {
      *
      * @return height of claw as read by the encoder, in ticks
      */
-    public double getCurrentHeight() {
+    public int getCurrentHeight() {
         return elevatorTalon.getSelectedSensorPosition(RobotMap.PID.PRIMARY_PID_LOOP);
     }
 }
